deployment/api/requirements.txt
# Core API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# ML Libraries
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3
lightgbm==4.1.0
xgboost==2.0.3

# SHAP for interpretability
shap==0.42.1

# GCP
google-cloud-storage==2.10.0

# Utilities
python-multipart==0.0.6
python-json-logger==2.0.7
requests==2.31.0
